+++ CHANGELOG.md
@@ -1,6 +1,26 @@
 CHANGELOG
 =========
 
+0.43.1
+------
+- (Experimental) Added support for Sixel graphics in the preview window
+  ```sh
+  # 1. $FZF_PREVIEW_WIDTH and $FZF_PREVIEW_HEIGHT will be set to the pixel width
+  #    and height of the preview window
+  # 2. Special preview window flag 'clear' is added to always completely
+  #    erase the preview window. This is similar to https://github.com/vifm/vifm/issues/588.
+  fzf --preview='
+    if file --mime-type {} | grep -qvF image/; then
+      bat --color=always {}
+    elif [[ -n $FZF_PREVIEW_WIDTH ]]; then
+      convert {} -resize ${FZF_PREVIEW_WIDTH}x${FZF_PREVIEW_HEIGHT} sixel:-
+    else
+      echo "Cannot display image data (unsupported platform)"
+    fi
+  ' --preview-window clear
+  ```
+- Bug fixes
+
 0.43.0
 ------
 - (Experimental) Added support for Kitty image protocol in the preview window